### 2019-12-20 #1974 @rhagenson
I noticed I placed parentheses around citations which then resulted in double parentheses on the proof. I will submit a fix for that in just a minute and regenerate the pdf.
