### 2019-12-20 #1974 @rhagenson
@arfon Quick question related to the formatting flubs I am currently fixing. Is there a way to run whedon locally to check formatting before submission? I did not see a way to do this, but that certainly does not mean they do not exist somewhere.
