### 2019-12-20 b0fc265
Fixing citation formatting
