### 2020-01-24 #1974 @marcosvital
Hi, @rhagenson , sorry about the delayed reply, I was still organizing myself after some weeks out of office.

If everything seems to be ok with the article proof, we can move on to assigning reviewers for you submission. We usually ask the authors for suggestions, so let me know if you have anyone in mind that could do this.

You can also check this [list of people](https://bit.ly/joss-reviewers) that have agreed to review for JOSS, and see  and may be suitable for this submission.

And let me know if you need any help finding reviewers if you don't have anyone in mind.
