### 2020-01-24 #1974 @rhagenson
@marcosvital, looking at the list of reviewers, the folks I found via a search for Go and visualization experience are:

@ djmitche
@ terrytangyuan
@ rabdill

Spaces after the @ to let you make the decision who to reach out to and in what order.
