### 2020-02-06 #1974 @marcosvital
Hi, @rhagenson . thanks for pointing potential reviewers for your submission, we can move on and ask them to review.

Dear @djmitche, @terrytangyuan and @rabdill :
Would you be willing to review this submission for JOSS?

This is still a pre-review, once there are enough reviewers, we will move on to a new issue where the review will take place.

If you are not able, please let us know if you have recommendations for other reviewers.
