### 2020-02-11 #1974 @marcosvital
Thank you @rabdill and @djmitche! I'll wait a few more to see if @terrytangyuan can also join us, but we will move soon to review, ok? This will happen in a new issue, and if yo haven't reviewed for JOSS before, you can take a look [at this guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
