### 2020-02-11 #1974 @arfon
> I am now an editor of JOSS so likely won't have time to participate in reviews. Thanks for checking though. BTW, @arfon could you remove me from the reviewers spreadsheet?

Done!
