+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-Given the relatedness of a set of individuals as calculated by tools such a ML-Relate [@ML-Relate], in the past the compiling biologist had to manually draw the resulting pedigree chart via manual reasoning on inferred connections --`relped` automates this process. By using a combination of relatedness values (with optional normalization) and codified entries, `relped` allows incremental refinement of the resulting pedigree network by replacing values with codes as familial determinations are definitively called. As well, there are options to include parentage and/or demographics information to better match the typical appearance of a standard pedigree chart. With parentage information provided, `relped` can differentiate between parent-offspring and full-sibling (both $\approx 0.5$ related) relationships. Pedigree network connections made between individuals are consistent between runs of `relped`; therefore, the pedigree network can be redrawn by repeatedly running `relped` until the pedigree network is visually satisfactory. Given the findings of @Staples2016, by default relationships beyond the 9th-level (e.g., 4th cousins) are not mapped. Any individuals related beyond this cutoff are left unmapped, but a list of these IDs can be saved.
+Given the relatedness of a set of individuals as calculated by tools such a ML-Relate [@ML-Relate], in the past the compiling biologist had to manually draw the resulting pedigree chart via manual reasoning on inferred connections -- `relped` automates this process. By using a combination of relatedness values (with optional normalization) and codified entries, `relped` allows incremental refinement of the resulting pedigree network by replacing values with codes as familial determinations are definitively called. As well, there are options to include parentage and/or demographics information to better match the typical appearance of a standard pedigree chart. With parentage information provided, `relped` can differentiate between parent-offspring and full-sibling (both $\approx 0.5$ related) relationships. Pedigree network connections made between individuals are consistent between runs of `relped`; therefore, the pedigree network can be redrawn by repeatedly running `relped` until the pedigree network is visually satisfactory. Given the findings of @Staples2016, by default relationships beyond the 9th-level (e.g., 4th cousins) are not mapped. Any individuals related beyond this cutoff are left unmapped, but a list of these IDs can be saved.
 
 # Statement of Need
 
