### 2020-02-27 #2124 @whedon
[REVIEW]: relped: Build Relatedness Pedigrees
**Submitting author:** @RHagenson (<a href="http://orcid.org/0000-0001-9750-1925">Ryan Hagenson</a>)
**Repository:** <a href="https://github.com/rhagenson/relped" target ="_blank">https://github.com/rhagenson/relped</a>
**Version:** v1.0.0
