### 2020-03-02 #2124 @marcosvital
Dear @djmitche and @rabdill, thank you very much for accepting review this submission for JOSS.

Please check the instructions and checklists above (I see that @djmitche already started), and let me know if you need any assistance.
You can also tag @rhagenson if you need to ask specific questions about the submission.
