### 2020-03-11 #2124 @djmitche
I'm having a bit of trouble figuring out what this does, as someone not familiar with this kind of analysis.  The README is very clear about input formats, but for example doesn't explain what "relatedness" means.  Is there another resource that could help me understand this sort of thing?
