### 2020-03-11 #2124 @rhagenson
@djmitche Great question and a bit of oversight on my part assuming relatedness is more widely known. In short, relatedness is the measure of how much two individuals overlap genetically -- ranging from 0 (no relation) to 1 (clones). `relped` began its lifecycle as a quasi-extension to a tool called [ML-Relate](http://www.montana.edu/kalinowski/software/ml-relate/index.html), which is where the codes under [Relatedness](https://github.com/rhagenson/relped#relatedness) originated from.

Perhaps the best reference for understanding relatedness is ["Estimating relatedness using genetic markers"](https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1558-5646.1989.tb04226.x) by Queller and Goodnight.

Does that help? Do you believe I need a brief section explaining relatedness in the README?
