### 2020-03-11 #2124 @marcosvital
@djmitche, I think that would be nice to explain that on the manuscript's summary.

It's actually on JOSS guidelines, the summary should "have a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided".

So a quick and simple explanation on relatedness should be there - and perhaps a slightly more detailed explanation could fit in the README.
