### 2020-03-11 #2124 @rhagenson
@rabdill, @djmitche, and @marcosvital I will add a briefing on relatedness to the paper and README to frame `relped`.

@rabdill, specifically, when you say "there was very little background for how this could be applied" do you mean you had difficulties identifying the usefulness of `relped`, or where it fits in a hypothetical project?

If it would be helpful, I can put a few sentences in the paper about where the methodology was required. We developed `relped` to build the pedigree network of hundreds of captive animals to inform national breeding programs. However, this is not the only use of `relped` so I did not force this context into either the paper or documentation -- where it might distract from the form and function of the software itself.
