### 2020-03-11 RHagenson/relped#40 @rhagenson
Review: Add Brief on Relatedness
The question arose during review (https://github.com/openjournals/joss-reviews/issues/2124) that relatedness, its measure and purpose, are not discussed in either the manuscript, nor the documentation. A brief must be added in both places:

- [x] Manuscript: add paragraph on relatedness, its measure and purpose
- [x] Documentation: add paragraph on relatedness, its measure and purpose
