### 2020-03-11 91594b6
Less verbose definition of relatedness
