### 2020-03-12 RHagenson/relped#40 @rhagenson
Addition to manuscript in 85571af32369276a1e5fcd8bcdb518db6aa595ac
Addition to documentation in 0ff3d7a93873aab00a11c4b3e498add57421a2ee

@ThatLionLady, you should have Pull Request (PR) assigned to you now. If you go to https://github.com/rhagenson/relped/pull/41/files you will see the side-by-side of the two versions. If you approve of the changes, you should be able to mark them as approved. Otherwise, make any comments in the PR thread and I will make edits accordingly until you approve then we will merge the changes.
