### 2020-03-12 6e7fdf2
Add Queller and Goodnight citation
