+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-Relatedness is a measure of shared genes and ranges from 0.0 (no relation) to 1.0 (exact clones). By numerically tracing genetics through a family tree, or pedigree chart, we can establish: 0.5 relatedness between parent-offspring (50% of the offspring's genes are from one parent), ~0.5 relatedness of full-siblings (50% of genes from each parent individually, but not the same 50% to each offspring), ~0.25 relatedness of half-siblings (50% of genes from one parent individually, but not the same 50%) -- this trend toward 0.0 continues the more distantly related two individuals are genetically.
+Relatedness is a measure of shared genes and ranges from 0.0 (no relation) to 1.0 (exact clones) [@Queller1989]. By numerically tracing genetics through a family tree, or pedigree chart, we can establish: 0.5 relatedness between parent-offspring (50% of the offspring's genes are from one parent), ~0.5 relatedness of full-siblings (50% of genes from each parent individually, but not the same 50% to each offspring), ~0.25 relatedness of half-siblings (50% of genes from one parent individually, but not the same 50%) -- this trend toward 0.0 continues the more distantly related two individuals are genetically.
 
 Given the relatedness of a set of individuals as calculated by tools such a ML-Relate [@ML-Relate], in the past the compiling biologist had to manually draw the resulting pedigree chart via manual reasoning on inferred connections -- `relped` automates this process. By using a combination of relatedness values (with optional normalization) and codified entries, `relped` allows incremental refinement of the resulting pedigree network by replacing values with codes as familial determinations are definitively called. As well, there are options to include parentage and/or demographics information to better match the typical appearance of a standard pedigree chart. With parentage information provided, `relped` can differentiate between parent-offspring and full-sibling (both $\approx 0.5$ related) relationships. Pedigree network connections made between individuals are consistent between runs of `relped`; therefore, the pedigree network can be redrawn by repeatedly running `relped` until the pedigree network is visually satisfactory. Given the findings of @Staples2016, by default relationships beyond the 9th-level (e.g., 4th cousins) are not mapped. Any individuals related beyond this cutoff are left unmapped, but a list of these IDs can be saved.
 
