### 2020-03-12 7aeaa45
Add brief statements on relatedness (#41)
* Add relatedness paragraph to Summary

* Less verbose definition of relatedness

* Add brief on relatedness to contextualize relped

* Add Queller and Goodnight citation
