### 2020-03-14 #2124 @rhagenson
As far as I know, I am as available as usual so no immediate need to pause.
However, have no expectation of the same reality among this submission's
editor and reviewers.

Please be safe and protect yourselves; this submission can wait for
everyone to be happy and healthy during these times. =)
