### 2020-03-16 #2124 @rabdill
Hi! Working my way through the rest of the review now, thank you for your patience. I have a few minor notes, but I ran into something that might be a bigger issue, so I'll just post that until I can make more progress.

I'm not clear on whether this is a problem or not, but I noticed that the **inferred connections are not consistent between runs**. I realize the layout of the diagrams may change, but the readme states that "all connections will be the same between runs," and not all of the differences I found could be fixed by just adding the "undirected" option, so I wanted to bring this up just in case.

I generated four diagrams using identical inputs with
```sh
relped build --relatedness="example-data/relatedness-nums-and-codes.csv" \
  --parentage="example-data/parentage.csv" --output "asdf.dot"
```

It generated output exactly as described, and I **included the figures below** with red circles around a few discrepancies I noticed:
* iteration 1: connection from M2 to F3
  * iteration 2: M2 connected to F3, **then *another* connection from F3 back to M2** (should this happen?) not present in the other iterations
  * iterations 3 and 4: no connection between M2 and F3, but **both have a connection from F7 to F3 instead**.
* iteration 1 and 3: connection from O4 to O11
  * iteration 2: no connection
  * iteration 4: O11 connected to O4 in the opposite direction
* iteration 1: no connection between M1 and F8
  * iterations 2 and 3: connection from F8 to M1
  * iteration 4: M1 connected to F8 in the opposite direction
* iterations 1 and 2: connection from F2 to M2
  * iterations 3 and 4: connection from M2 to F2 in the opposite direction

I thought maybe this would go away with more data, so I tried running it again including the demographics file, but I found several inconsistent connections between the first 2 iterations there as well. @rhagenson, would you mind clarifying if this is expected behavior? Please let me know if it'd be helpful to have this in a different format too; I'd be happy to send it some other way. Thanks.

![1](https://user-images.githubusercontent.com/6011419/76791988-0381a100-6790-11ea-8dbf-1b298e50efc1.png)
![2](https://user-images.githubusercontent.com/6011419/76791670-645ca980-678f-11ea-9953-bc5a858cd014.png)
![3](https://user-images.githubusercontent.com/6011419/76791676-6888c700-678f-11ea-8b06-69c5cf8938ad.png)
![4](https://user-images.githubusercontent.com/6011419/76791683-6d4d7b00-678f-11ea-9ce9-be223fcda081.png)
