### 2020-03-16 #2124 @rhagenson
Loops, such as in iteration 2 should be removed already. That one is going to take some head work to solve as there is already a cycle removal step.
