### 2020-03-16 #2124 @rhagenson
> Loops, such as in iteration 2 should be removed already. That one is going to take some head work to solve as there is already a cycle removal step.

I think I know what I did here. The step did exist, but then at a later stage of development I removed it as the (mistakenly) thought the improved pruning covered it already so removed what was thought of as a unnecessary repetition of work.
