### 2020-03-16 #2124 @rhagenson
One the F2 to M2 disconnect. See the above comment on M1 and F8.

> Ensuring any codified entries cannot be dropped (as I would expect from the program as a user) will fix this issue
