### 2020-03-16 #2124 @rhagenson
Overall, @rabdill, it seems you IDed some regressions in the program logic which predated effective testing for such regressions.

The two remedies that should fix all the ID problems are:

+ Fixing the loop-removal procedure to ensure no loops between different unknown/inferred individuals among known individuals
+ Specially mark any codified relationship and ensure those relationships cannot be removed

Thank you for catching these issues!
