### 2020-03-16 RHagenson/relped#42 @rhagenson
Remove loops
As identified during review by https://github.com/openjournals/joss-reviews/issues/2124#issuecomment-599716247, some runs of `relped` still produce loops through different unknown individuals.

From that thread:
![](https://user-images.githubusercontent.com/6011419/76791670-645ca980-678f-11ea-9953-bc5a858cd014.png)

On the right-hand side, we can see a path M2->F3 and a corresponding reverse path F3->M2.

There is already a procedure meant to remove such loops, but it appears to not remove them consistently.
