### 2020-03-16 RHagenson/relped#43 @rhagenson
Ensure codified entries always remain
As identified during review by https://github.com/openjournals/joss-reviews/issues/2124#issuecomment-599716247, `relped` is currently dropping some codified entries such as half-siblings. I am uninterested in precisely "why" these relationships are being dropped as I suspect it is a case of pruning and the order that pruning happens to occur in for a particular iteration.\*

What is important here is to match user expectation where if they have entered a code, then that connection should **always** exist in the final pedigree network.

\*: If two paths have the same weight because they were both drawn from the same code `relped` during pruning selects the "highest weight" path then drops the others even if equally weighted.
