### 2020-03-20 #2124 @djmitche
I feel like I'm playing catch-up - thanks for your detailed analysis @rabdill!  A few minor notes:

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I think 2 and 3 are covered, but 1 is not.  [GitHub has suggestsions](https://github.com/rhagenson/relped/community) for both a contributing guide and a CoC.  The CoC probably isn't going to see any use at this stage, but the few times I have had to call on a CoC I've been glad it was already in place, so I'd suggest creating one.

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I see that there is example data in the repo, but it's not clear which files belong together.  Also, the only example usage is in the multiple-runs example.  Is it expected that the software is always used in that mode, or would it make sense to give a single-run example that uses data from the example-data directory?
