### 2020-03-24 #2124 @rhagenson
@djmitche

I can add a CONTRIBUTING doc and agree that a CoC is probably not going to be used at this stage, but could be useful to have in place well before it needs to be used.

The `example-data` files were named according to what option in the CLI they correspond to, therefore all the `relatedness-*` files go to `--relatedness` and same for `parentage*` and `demographics*`. As for multi versus single run, I would expect multi-run is the manner most users will run it in; I only have some much control over how GraphViz will layout a particular `relped` output so multi-run allows for rapidly building what should be the same semantic graph where connections are consistent, but layout changes (Rich pointed out this was not always the case). I will still add a single run example just above the multi run example as it shouldn't distract from usage by any means so is harmless at worst and helpful at best.
