### 2020-03-24 #2124 @djmitche
If the multi-run option is the normal course, it's fine to list only that example.  Perhaps a small rewording of the README would help others not wonder the same thing I did :)
