### 2020-03-29 #2124 @djmitche
I'm sorry I've been slow to get back to this -- I really have no excuse.  At this point I'm just waiting on CONTRIBUTING.

One thing in the paper that leaves me with a lingering uncertainty is "As more information is provided, connections are better resolved."  Perhaps this is my ignorance of the subject matter, but given a body of relatedness information that does not uniquely identify a single pedigree, what does it mean to be "poorly resolved"?  Or to ask another way, what is the prior state on which "better resolved' is based?  I can imagine that a user of the software would want to know concretely, would it ever draw an edge between two individuals that would later be removed given additional (non-contradictory) information?

If that's already known to someone doing this kind of work, then please consider it a question for my own curiosity (and don't feel obliged to answer it!)
