### 2020-03-31 70181ab
Add CONTRIBUTING doc
