### 2020-03-31 #2124 @rhagenson
CONTRIBUTING has been [added](https://github.com/rhagenson/relped/blob/master/CONTRIBUTING.md)
