### 2020-04-03 #2124 @djmitche
Thanks for the explanation.  The key point for my review is "wording that is understood by those doing the work already" :)
