### 2020-04-13 #2124 @rabdill
Hi @arfon. I think my review is technically the missing piece. I'm not sure how this works within the JOSS review process, but I was hoping to wait for the completion of two issues that popped up [during review](https://github.com/openjournals/joss-reviews/issues/2124#issuecomment-599716247): https://github.com/rhagenson/relped/issues/42 https://github.com/rhagenson/relped/issues/43

In short, one of the stated expectations of the software is that its conclusions are consistent across runs, but there are some cases where that isn't the case, and identical inputs can result in different outputs. I believe @rhagenson is working on it. Based on the functional claims of the software, I'm not comfortable checking the "Functionality" box until those things are fixed. I suppose an alternative would be to change the documentation to reflect a nondeterministic implementation, but it's my impression this would be a significant step backwards for the software itself.
