### 2020-04-21 ee9ca03
Add CoC doc
rm references to Pony
