### 2020-10-25 #2124 @arfon
> I agree that it should not be moved forward until I address the issues Rich noted. I apologize about any delays; I was correct that I am as available as ever, but that availability has not translated into productivity. My plan is to make significant progress on the aforementioned issues this week. Thank you to each of you for your patience with me.

@rhagenson@ @marcosvital - it looks like this submission is *very* close to crossing the line here. Is there anything we can do to help you wrap this up :soon: ?

> Well, so much for that idea...

@djmitche apologies for this. @whedon 'helpfully' goes around looking for missing assignments on JOSS issues every 24 hours so even if you unassigned yourself, you'll get reassigned by Whedon.
