### 2020-11-02 #2124 @marcosvital
Hi, @rhagenson! Hope everything is fine with you.
Did you made any progress on the two issues? Let us know if you have any news.
