### 2020-11-12 #2124 @rhagenson
@marcosvital, I am still alive. I have many reasons for the delay, but the most informative is that in May I rapidly moved from my work in conservation (which this submission was a part of) to being a dedicated COVID-19 data analyst with my local health department. I have spent the last few months consuming COVID-19 literature, responding to outbreaks, and building computational scaffolding to help get us through this pandemic. I have no intention on abandoning this work, but I also cannot say for certain how soon I will have the time and energy to resolve the final two issues.

Context, I am dedicated to the largest city affecting the blue line below ([source](https://covid.cdc.gov/covid-data-tracker/#compare-trends_newcasesper100k)):
![Screenshot from 2020-11-12 12-22-51](https://user-images.githubusercontent.com/5414719/98980145-d71a5200-24e1-11eb-9f67-2e6685fa3e88.png)
