### 2020-11-18 #2124 @marcosvital
Hi, @rhagenson! Thank you for the update, I'm glad to know that you are fine.
Since you are not abandoning the project, we will keep this submission as paused, and wait for you.

@djmitche and @rabdill, thank you so much for all the time you dedicated into reviewing this submission. This is probably going to be the last notification for a while, until @rhagenson get to work on this again. When that happens, I can make the final checks if you happen to not be available anymore.
