### 2021-09-27 #2124 @arfon
@rhagenson – I appreciate that your attention is on other (very important) work, but JOSS does not have an open-ended commitment to you to wait for your updates (this submission is coming up to 2 years old now). As such, I'm going to go ahead and close this review. Should you find time to make these updates in the future we'd welcome a resubmission.

@djmitche @rabdill @marcosvital – many thanks for all of your efforts here.
